+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## Unreleased
 * Add prepaid_reloadable from bin data in credit card responses
+* Add support for `PayPalPaymentResource` requests
 
 ## 3.39.0
 * Add support for creating and updating PayPal customer session